requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.5
playwright==1.48.0
python-dotenv==1.0.1
